requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.115.12
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psutil==7.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
redis==6.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
